qcacmn: update average rssi in scan entry
If elapsed time since last rssi update for a particular
scan entry is smaller than a thresold, calculate a
running average of the RSSI values. Otherwise last RSSI
is more representive of the signal strength.

CRs-Fixed: 1095299

+++ umac/scan/dispatcher/inc/wlan_scan_public_structs.h
@@ -247,7 +247,7 @@ struct scan_cache_entry {
 	struct wlan_ssid ssid;
 	uint16_t seq_num;
 	enum wlan_phymode phy_mode;
-	int8_t avg_rssi;
+	int32_t avg_rssi;
 	int8_t rssi_raw;
 	uint16_t bcn_int;
 	union wlan_capability cap_info;
